turning sha1dir into a library
